+++ src/main/java/org/sonar/plugins/delphi/antlr/ast/node/FieldSectionNode.java
@@ -1,6 +1,5 @@
 package org.sonar.plugins.delphi.antlr.ast.node;
 
-import java.util.List;
 import org.antlr.runtime.Token;
 import org.sonar.plugins.delphi.antlr.ast.visitors.DelphiParserVisitor;
 
@@ -21,9 +20,5 @@ public final class FieldSectionNode extends DelphiNode implements Visibility {
   @Override
   public VisibilityType getVisibility() {
     return ((VisibilitySectionNode) jjtGetParent()).getVisibility();
-  }
-
-  public List<FieldDeclarationNode> getFieldDeclarations() {
-    return findChildrenOfType(FieldDeclarationNode.class);
   }
 }