+++ src/main/java/org/sonar/plugins/delphi/symbol/declaration/UnitNameDeclaration.java
@@ -1,5 +1,7 @@
 package org.sonar.plugins.delphi.symbol.declaration;
 
+import java.nio.file.Path;
+import java.util.Objects;
 import org.sonar.plugins.delphi.antlr.ast.node.FileHeaderNode;
 import org.sonar.plugins.delphi.symbol.scope.FileScope;
 
@@ -7,12 +9,14 @@ public final class UnitNameDeclaration extends QualifiedDelphiNameDeclaration {
   public static final String UNKNOWN_UNIT = "<unknown unit>";
   private final FileScope unitScope;
   private final String namespace;
+  private final Path path;
   private int hashCode;
 
   public UnitNameDeclaration(FileHeaderNode node, FileScope unitScope) {
     super(node.getNameNode(), unitScope);
     this.unitScope = unitScope;
     this.namespace = node.getNamespace();
+    this.path = Path.of(node.getASTTree().getFileName());
   }
 
   public FileScope getUnitScope() {
@@ -21,6 +25,10 @@ public final class UnitNameDeclaration extends QualifiedDelphiNameDeclaration {
 
   public String getNamespace() {
     return namespace;
+  }
+
+  public Path getPath() {
+    return path;
   }
 
   @Override
@@ -37,13 +45,13 @@ public final class UnitNameDeclaration extends QualifiedDelphiNameDeclaration {
       return false;
     }
     UnitNameDeclaration that = (UnitNameDeclaration) o;
-    return getImage().equalsIgnoreCase(that.getImage());
+    return getImage().equalsIgnoreCase(that.getImage()) && path.equals(that.path);
   }
 
   @Override
   public int hashCode() {
     if (hashCode == 0) {
-      hashCode = getImage().toLowerCase().hashCode();
+      hashCode = Objects.hash(getImage().toLowerCase(), path);
     }
     return hashCode;
   }