+++ src/test/resources/org/sonar/plugins/delphi/symbol/overloads/imports/StringFoo.pas
@@ -1,0 +1,16 @@
+unit StringFoo;
+
+{This is a sample Delphi file.}
+
+interface
+
+procedure Foo(Bar: String); overload;
+
+implementation
+
+procedure Foo(Bar: String);
+begin
+  // Do nothing
+end;
+
+end.